+++ src/java.base/share/classes/jdk/internal/foreign/NativeMemorySegmentImpl.java
@@ -31,20 +31,13 @@ import java.nio.ByteBuffer;
 import java.util.Optional;
 
 import jdk.internal.misc.Unsafe;
-import jdk.internal.misc.VM;
 import jdk.internal.vm.annotation.ForceInline;
 
 /**
  * Implementation for native memory segments. A native memory segment is essentially a wrapper around
  * a native long address.
  */
-public sealed class NativeMemorySegmentImpl extends AbstractMemorySegmentImpl permits MappedMemorySegmentImpl {
-
-    private static final Unsafe UNSAFE = Unsafe.getUnsafe();
-
-    // The maximum alignment supported by malloc - typically 16 bytes on
-    // 64-bit platforms and 8 bytes on 32-bit platforms.
-    private static final long MAX_MALLOC_ALIGN = Unsafe.ADDRESS_SIZE == 4 ? 8 : 16;
+sealed class NativeMemorySegmentImpl extends AbstractMemorySegmentImpl permits MappedMemorySegmentImpl {
 
     final long min;
 
@@ -109,73 +102,5 @@ public sealed class NativeMemorySegmentImpl extends AbstractMemorySegmentImpl pe
     @Override
     public long maxAlignMask() {
         return 0;
-    }
-
-    // factories
-
-    public static MemorySegment makeNativeSegmentNoZeroing(long byteSize, long byteAlignment, MemorySessionImpl sessionImpl,
-                                                  boolean shouldReserve) {
-        sessionImpl.checkValidState();
-        if (VM.isDirectMemoryPageAligned()) {
-            byteAlignment = Math.max(byteAlignment, NIO_ACCESS.pageSize());
-        }
-        long alignedSize = Math.max(1L, byteAlignment > MAX_MALLOC_ALIGN ?
-                byteSize + (byteAlignment - 1) :
-                byteSize);
-
-        if (shouldReserve) {
-            NIO_ACCESS.reserveMemory(alignedSize, byteSize);
-        }
-
-        long buf = allocateMemoryWrapper(alignedSize);
-        long alignedBuf = Utils.alignUp(buf, byteAlignment);
-        AbstractMemorySegmentImpl segment = new NativeMemorySegmentImpl(buf, alignedSize,
-                false, sessionImpl);
-        sessionImpl.addOrCleanupIfFail(new MemorySessionImpl.ResourceList.ResourceCleanup() {
-            @Override
-            public void cleanup() {
-                UNSAFE.freeMemory(buf);
-                if (shouldReserve) {
-                    NIO_ACCESS.unreserveMemory(alignedSize, byteSize);
-                }
-            }
-        });
-        if (alignedSize != byteSize) {
-            long delta = alignedBuf - buf;
-            segment = segment.asSlice(delta, byteSize);
-        }
-        return segment;
-    }
-
-    private static long allocateMemoryWrapper(long size) {
-        try {
-            return UNSAFE.allocateMemory(size);
-        } catch (IllegalArgumentException ex) {
-            throw new OutOfMemoryError();
-        }
-    }
-
-    // Unsafe native segment factories. These are used by the implementation code, to skip the sanity checks
-    // associated with MemorySegment::ofAddress.
-
-    @ForceInline
-    public static MemorySegment makeNativeSegmentUnchecked(long min, long byteSize, MemorySessionImpl sessionImpl, Runnable action) {
-        if (action == null) {
-            sessionImpl.checkValidState();
-        } else {
-            sessionImpl.addCloseAction(action);
-        }
-        return new NativeMemorySegmentImpl(min, byteSize, false, sessionImpl);
-    }
-
-    @ForceInline
-    public static MemorySegment makeNativeSegmentUnchecked(long min, long byteSize, MemorySessionImpl sessionImpl) {
-        sessionImpl.checkValidState();
-        return new NativeMemorySegmentImpl(min, byteSize, false, sessionImpl);
-    }
-
-    @ForceInline
-    public static MemorySegment makeNativeSegmentUnchecked(long min, long byteSize) {
-        return new NativeMemorySegmentImpl(min, byteSize, false, new GlobalSession(null));
     }
 }
